services/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncpg>=0.30.0
openai>=1.0.0
sse-starlette>=2.1.0
pydantic-ai>=0.7.4
supabase>=2.0.0
httpx>=0.25.0
docling>=2.55.0
numpy>=2.0.2
hf-xet>=1.1.8
